When we found error in dr. solution imlpement if (image instanceof GrayImage)

+++ src/interfaces/ImageProcessing.java
@@ -4,9 +4,9 @@ import java.util.*;
 
 
 public interface ImageProcessing {
-    public abstract Image thresholding (Image image, byte thresholdValue);
+    public abstract Image thresholding (Image image, int thresholdValue);
     public abstract Image toDouble (Image image);
-    public abstract Map<Byte, Integer> Histogram (Image image);
+    public abstract Map<Integer, Integer> Histogram (Image image);
     public abstract int MatchPattern (Image subimage, Image image);
     
 }
